Payment process refactor
